Remove benchmark-caliper again and switch EngineType to an interface.
