--- benchmark-caliper/src/main/AndroidManifest.xml
@@ -1,6 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="org.conscrypt.benchmarks.caliper">
-
-    <application/>
-
-</manifest>
